move service back into telemetry.go
Copilot-Session: 9452b3e7-02d8-4d4d-b2db-ab18b35657c1
