+++ internal/telemetry/telemetry.go
@@ -4,17 +4,24 @@ package telemetry
 
 import (
 	"bytes"
+	"encoding/binary"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
+	"maps"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"runtime"
 	"slices"
+	"strconv"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/cli/cli/v2/internal/config"
+	"github.com/cli/cli/v2/internal/gh/ghtelemetry"
 	"github.com/cli/cli/v2/pkg/jsoncolor"
 	"github.com/google/uuid"
 	"github.com/mgutz/ansi"
@@ -138,6 +145,28 @@ func ParseTelemetryState(configValue string) TelemetryState {
 	return Enabled
 }
 
+type serviceOptions struct {
+	additionalDimensions ghtelemetry.Dimensions
+	sampleRate           int
+}
+
+type serviceOption func(*serviceOptions)
+
+// WithAdditionalCommonDimensions sets dimensions shared by every invocation event.
+func WithAdditionalCommonDimensions(dimensions ghtelemetry.Dimensions) serviceOption {
+	return func(options *serviceOptions) {
+		maps.Copy(options.additionalDimensions, dimensions)
+	}
+}
+
+// WithSampleRate selects invocation-wide sampling. Rates 0 and 100 retain all
+// events; rates between them select a percentage using the invocation ID.
+func WithSampleRate(rate int) serviceOption {
+	return func(options *serviceOptions) {
+		options.sampleRate = rate
+	}
+}
+
 // LogFlusher returns a flush function that writes telemetry payloads to the provided log writer. This is used for the "log" telemetry mode, which is intended for debugging and development.
 // When there are no events to report (for example the command opted out of telemetry, the user is on GHES, or no events were recorded), a "Telemetry payload: none" marker is written so that the absence of events is observable.
 var LogFlusher = func(log io.Writer, colorEnabled bool) func(payload SendTelemetryPayload) {
@@ -177,6 +206,195 @@ var GitHubFlusher = func(executable string) func(payload SendTelemetryPayload) {
 			return
 		}
 		SpawnSendTelemetry(executable, payload)
+	}
+}
+
+// NewService creates a telemetry service using send to deliver its completed payload.
+func NewService(send func(SendTelemetryPayload), opts ...serviceOption) *Service {
+	options := serviceOptions{
+		additionalDimensions: make(ghtelemetry.Dimensions),
+	}
+	for _, opt := range opts {
+		opt(&options)
+	}
+
+	deviceID, err := deviceIDFunc()
+	if err != nil {
+		deviceID = "<unknown>"
+	}
+	invocationID := uuid.NewString()
+	commonDimensions := ghtelemetry.Dimensions{
+		"device_id":     deviceID,
+		"invocation_id": invocationID,
+		"os":            runtime.GOOS,
+		"architecture":  runtime.GOARCH,
+	}
+	maps.Copy(commonDimensions, options.additionalDimensions)
+
+	hash := uuid.NewSHA1(uuid.Nil, []byte(invocationID))
+	sampleBucket := byte(binary.BigEndian.Uint32(hash[:4]) % 100)
+
+	return &Service{
+		send:             send,
+		commonDimensions: commonDimensions,
+		sampleRate:       options.sampleRate,
+		sampleBucket:     sampleBucket,
+	}
+}
+
+type invocationEvent struct {
+	event      ghtelemetry.Event
+	recordedAt time.Time
+}
+
+var (
+	_ ghtelemetry.Service = (*Service)(nil)
+	_ ghtelemetry.Service = (*NoOpService)(nil)
+)
+
+// Service records telemetry facts and reporting policy for one command execution.
+// Finish must run after command execution to send the completed payload.
+type Service struct {
+	mu               sync.Mutex
+	send             func(SendTelemetryPayload)
+	commonDimensions ghtelemetry.Dimensions
+	sampleRate       int
+	sampleBucket     byte
+	events           []*invocationEvent
+	disabled         bool
+	finished         bool
+}
+
+type pendingEvent struct {
+	service  *Service
+	recorded *invocationEvent
+}
+
+// Disable suppresses all events in the invocation, including already recorded
+// events. It must be called before Finish.
+func (s *Service) Disable() {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	s.disabled = true
+}
+
+// Record copies a complete event into the service.
+// Recording after Finish has no effect.
+func (s *Service) Record(event ghtelemetry.Event) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if s.finished {
+		return
+	}
+	s.events = append(s.events, &invocationEvent{
+		event:      cloneEvent(event),
+		recordedAt: time.Now(),
+	})
+}
+
+// Begin copies an event's initial facts and returns a handle for adding
+// facts until Finish. Events begun after Finish are not recorded.
+func (s *Service) Begin(event ghtelemetry.Event) ghtelemetry.PendingEvent {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if s.finished {
+		return noOpPendingEvent{}
+	}
+	recorded := &invocationEvent{
+		event:      cloneEvent(event),
+		recordedAt: time.Now(),
+	}
+	s.events = append(s.events, recorded)
+	return &pendingEvent{service: s, recorded: recorded}
+}
+
+func (p *pendingEvent) SetDimensions(dimensions ghtelemetry.Dimensions) {
+	p.service.mu.Lock()
+	defer p.service.mu.Unlock()
+
+	if p.service.finished {
+		return
+	}
+	if p.recorded.event.Dimensions == nil {
+		p.recorded.event.Dimensions = make(ghtelemetry.Dimensions)
+	}
+	maps.Copy(p.recorded.event.Dimensions, dimensions)
+}
+
+func (p *pendingEvent) SetMeasures(measures ghtelemetry.Measures) {
+	p.service.mu.Lock()
+	defer p.service.mu.Unlock()
+
+	if p.service.finished {
+		return
+	}
+	if p.recorded.event.Measures == nil {
+		p.recorded.event.Measures = make(ghtelemetry.Measures)
+	}
+	maps.Copy(p.recorded.event.Measures, measures)
+}
+
+// SetSampleRate selects the sampling policy for the whole invocation.
+// Changes after Finish have no effect.
+func (s *Service) SetSampleRate(rate int) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if s.finished {
+		return
+	}
+	s.sampleRate = rate
+	s.commonDimensions["sample_rate"] = strconv.Itoa(rate)
+}
+
+// Finish snapshots the recorded events once and sends their payload after releasing the lock.
+// Sampling and telemetry eligibility apply to immediate and pending events alike.
+func (s *Service) Finish() {
+	s.mu.Lock()
+
+	if s.finished {
+		s.mu.Unlock()
+		return
+	}
+	s.finished = true
+
+	if s.sampleRate > 0 && s.sampleRate < 100 && int(s.sampleBucket) >= s.sampleRate {
+		s.mu.Unlock()
+		return
+	}
+
+	events := s.events
+	if s.disabled {
+		events = nil
+	}
+
+	// Keep an empty payload so log mode can explain that no telemetry will be sent.
+	payload := SendTelemetryPayload{Events: make([]PayloadEvent, len(events))}
+	for index, recorded := range events {
+		dimensions := map[string]string{
+			"timestamp": recorded.recordedAt.UTC().Format("2006-01-02T15:04:05.000Z"),
+		}
+		maps.Copy(dimensions, s.commonDimensions)
+		maps.Copy(dimensions, recorded.event.Dimensions)
+		payload.Events[index] = PayloadEvent{
+			Type:       recorded.event.Type,
+			Dimensions: dimensions,
+			Measures:   maps.Clone(recorded.event.Measures),
+		}
+	}
+	s.mu.Unlock()
+
+	s.send(payload)
+}
+
+func cloneEvent(event ghtelemetry.Event) ghtelemetry.Event {
+	return ghtelemetry.Event{
+		Type:       event.Type,
+		Dimensions: maps.Clone(event.Dimensions),
+		Measures:   maps.Clone(event.Measures),
 	}
 }
 
@@ -262,3 +480,28 @@ func SpawnSendTelemetry(executable string, payload SendTelemetryPayload) {
 	// Release resources associated with the child process since we will never Wait for it.
 	_ = cmd.Process.Release()
 }
+
+type noOpPendingEvent struct{}
+
+func (noOpPendingEvent) SetDimensions(ghtelemetry.Dimensions) {}
+func (noOpPendingEvent) SetMeasures(ghtelemetry.Measures)     {}
+
+// NoOpService discards telemetry when collection is disabled.
+type NoOpService struct{}
+
+// Record discards the event.
+func (*NoOpService) Record(ghtelemetry.Event) {}
+
+// Begin returns an inert handle without retaining the event.
+func (*NoOpService) Begin(ghtelemetry.Event) ghtelemetry.PendingEvent {
+	return noOpPendingEvent{}
+}
+
+// Disable leaves telemetry disabled.
+func (*NoOpService) Disable() {}
+
+// SetSampleRate leaves telemetry disabled.
+func (*NoOpService) SetSampleRate(int) {}
+
+// Finish has no payload to complete.
+func (*NoOpService) Finish() {}
